Handle date parsing and formatting
